Don't call 'getShortDescription' when tracing is not enabled
The call to 'getShortDescription' is expensive. Currently, it's called
even if tracing is not enabled, which slows down input processing. Move
this into an 'if' statement, which will ensure that it's only called if
tracing is enabled.

If we find that this is still too expensive (slows down input processing
too much when tracing is enabled), we should consider getting rid of
this trace completely.

Bug: 395115521
Flag: EXEMPT bugfix
Test: none
Change-Id: Icbf2ea32e6ebc4c944d214c953b11a5d7a7cdede