Enable multiple subtypes in InputMethodServiceTest
Previously in testImeSwitchButtonClick we were checking if either the
menu was shown, or the IME is different (switched away). Currently this
would always be the second case as the actual system IME was still
enabled. However, we should not rely on this fact, but rather enable
addtional subtypes in our SimpleIME and check that we always end up
swiching the subtype.

This also ensures that the SimpleIME is always the currently running IME
at the end of each test.

Flag: EXEMPT testfix
Bug: 394328311
Test: atest InputMethodServiceTest#testImeSwitchButtonClick
  InputMethodServiceTest#testImeSwitchButtonLongClick
Change-Id: I16986c288565b060140a85090b5ba2fb4506bad4