Fix bouncer transitions when bouncer shown over hub
When the bouncer is shown over the glanceable hub, we keep the hub UI
visible underneath the bouncer and don't navigate away from it. This
means that if/when the bouncer navigates away to another state, it
should also close the hub. If we don't close the hub correctly, then it
will keep being displayed on top of other content.

Bug: 394915563
Test: atest FromPrimaryBouncerTransitionInteractorTest
Test: atest FromAlternateBouncerTransitionInteractorTest
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I35a7009d6866870f50c3bd5d9b6878a540155a99