Removing IMS surface in InputMethodService#hideSoftInput after hiding the window
After hiding the IME, we formerly removed its surface via an explicit call in the ImeInsetsSourceConsumer, later in InsetsController, but was requested too early, as the animation was not finished at the time. Removing the surface should only happen, after the IME window was already hidden and can be done in InputMethodService#hideSoftInput.

Bug: 298172246
Flag: android.view.inputmethod.refactor_insets_controller
Test: atest com.android.inputmethodservice.InputMethodServiceTest#testSurfaceRemovedAfterHideSoftInput
Change-Id: I546327a5dfb570dcf3999f94e4556d08ac2a70b6