Update the expanded view bounds after config change
Note that this updates the bounds after a delay so there's a 
period of time where the bounds are incorrect after unfolding.
This delay exists today, and this change ensures that the
view isn't stuck with incorrect bounds.

The proper solution is to merge the unfold transition with the
task view transition, which is tracked in b/392893178

Flag: EXEMPT bug fix
Bug: 390050545
Test: manual
       - have a floating bubble
       - expand and show IME
       - unfold
       - observe expanded view bounds update correctly
       - fold and repeat

Change-Id: I4c2faa77937ed92986ec8c5715ef79faa994a8e9