Fix forced auto-grouping mark notifications as unseen issue
The force auto-grouping will update the notifications in the
NotifCollection, which makes such notifications unseen. There was an
existing parameter, fromSystem, in onEntryUpdated() method, however,
this parameter is not accurate enough for us to distinguish an update
from the system server(eg. auto-grouping) and a real update from App,
(because that is also made through system server).

This change replaced the fromSystem parameter with a Enum class:
UpdateSource, which distinguishes system server, app, and systemui
updates. We are not going to mark an existing notification unseen again
if it's updated by a source other than App.

Fix: 391443182
Test: RemoteInputCoordinatorTest, BubblesTest
Flag: com.android.server.notification.notification_minimalism
Change-Id: I8d95dc6c3878465071c7ed2cb1de28d7a5d6ec43