Adding roundness during pre-detachment.
The MagneticRowManagerImpl works closer with the
NotificationRoundnessManager to set a linearly changing roundness of
notifications as the user drags a swiped notification. The roundness
gets fully set to 1 when the notification is swiped past the magnetic
detach threshold.

Roundness gets reset when a notification is dismissed and when the snap
back animation from releasing it overshoots for the first time.

Test: MagneticRowManagerImplTest
Flag: com.android.systemui.magnetic_notification_swipes
Bug: 397942630

Change-Id: I25aca40073fe06c5fee618d1d10f26495dd0fcbd