[Notif redesign] Include imageMessageContainer in animation
A redesign change accidentally excluded the imageMessageContainer from the list of drawables used in the conversation animation. While it's true that it's now nullable where previously it was not (since it doesn't exist in the collapsed layout), we should still include it when not null.

Change-Id: Ia0e780f8de71e7dea6921bc3a9dcfa99e8fa9cb6
Fix: 397900637
Flag: android.app.notifications_redesign_templates
Test: NotificationConversationTemplateViewWrapperTest