Remove addFlow/TerminatingFlow
setFlow and setTerminatingFlow had been added in previous cl
to replace the addFlow methods. In practice, addFlow is very
unlikely to be used to add more than one edge to an event. In
these cases, there's a noticeable performance improvement when
using setFlow which avoids the call through to Pool of Flows.

Also fixed setFlow to accept long instead of int. The underlying
structs in the C SDK support uint64_t and this was just an
oversight.

Test: atest PerfettoTraceTest
Bug: 303199244
Flag: android.os.perfetto_sdk_tracing_v2
Change-Id: I3fdcbbd09921e865052bdca359a5ffb0cfe18900