Revert "Check sound Uri permission when creating a notification channel"
This reverts commit 94fc92cf20b897f016c3a9718edf40818de4a11b.

Reason for revert: CVE-2025-22420 A-337775777 is being removed from ASB#2025-04 due to an possible app compatibility regression. Partners are recommended to remove this fix and should not reference it in any public documentation. An updated fix will be required as part of a future Android Security Bulletin.

Pixel Impact: Internal testing found WhatsApp java crashes with this fix. Detected on Dogfood builds. We recommend Pixel removal if possible. Crash report here b/390020418

Change-Id: Iaf59e34577bed2eba8da30037a77d15e4f4bd690