Add ability to override the fits system windows property of onclick dialog.
The PreferenceScreen will display a dialog in the onClick handler.  Adding
a method that legs the consumer of the PreferenceScreen override the
setFitsSystemWindows property on the list that is shown in the dialog.

This is done to fix a bug seen with the voicemail number in the voicemail
activity in Telephony.  I don't know where else this is problematic;
it seems the impacted screen in Telephony has a PreferenceScreen nested in
a PreferenceScreen which appears to cause the issue.  The change is made
in this manner to explicitly limit the scope to instances which have
demonstrated the issue.

Flag: NONE bug fix.
Fixes: 394283757
Fixes: 398759149
Test: Manual test of the telephony voicemail screen to demonstrate that
edge to edge does not break the layout.

Change-Id: I2dd9ba2b43f8a14cc29111e08064ebfceb4548a6