Add task scale to the convert to bubble animation
When a task is dragged to a bubble, the task is scaled down.
Add the initial scale to the size change animation and include another
animator in the set to scale the task up from the drag scale to 1x
scale.
Apply the dragged task scale to the snapshot while it is not under the
task container and while the task is being moved to the bubble task view.
Pass along full task bounds along with a scale to the size change animation.
This way the size change animation will be animating from start bounds
with intial scale to end bounds.
Update to emphasized interpolator for the size change animation.

Bug: 396449031
Test: atest BubbleTransitionsTest
Test: manual, drag a task to bubble
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: I192c97aebe2632e6519d75ea9739e0037ca5484b