[Media] Output device chip padding for a11y.
Adds top, bottom, and end padding to extend the touch target of the
output device chip such that it's at least 48dp in height.

This changes the top row to be a Box and the top-level padding around
the entire card to be eliminated and replaced by padding around all
three rows.

Bug: 397989775
Test: manually verified in the compose gallery app that touching above,
to the right, or right below the chip also triggers the chip and shows
the ripple effect indication but only inside the chip bounds
Flag: EXEMPT code isn't used yet

Change-Id: I4dcc644e69be23df1a3641f9c69edea0011d24c6