Reset to page 0 if we stop listening
We will stop listening if always composed and not in the expanded QS, or
if we are in edit mode.

If we are not always composed, we will create a new PagerState the next
time we compose.

Test: manual, close shade in second page and reopen
Test: manual, go to second page, open edit mode and close
Test: atest QSFragmentComposeViewModelTest
Fixes: 401028991
Flag: com.android.systemui.qs_ui_refactor_compose_fragment
Flag: com.android.systemui.always_compose_qs_ui_fragment

Change-Id: I507ab4346b88ec921d94a436f0200724e9bf77c2