Fix SysUI crash with the bubble to fullscreen flag by updating the `provideDesktopPipTransitionObserver` method to check for the status of the flag.
When the bubble to fullscreen flag is enabled, DesktopTasksController is
created which depends on DesktopPipTransitionObserver.

The method `provideDesktopPipTransitionObserver` doesn't check for the
state of the bubble to fullscreen flag which might be empty resulting in
a crash on startup.

Bug: 401865824
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Test: manual (Enable bubble to fullscreen flag --> Verify SysUI doesn't
crash on startup
Change-Id: I3f712545f646e0e846f4dda3902b1bfa13986606

Change-Id: I3a122e9f9d661ec38f4bcf466588ca7075eef8f2