Confirm UID being reported
When JankTracking is first enabled the JankTracker object is passed the
apps UID. When stats are being reported from non platform libraries we
need to confirm the UID in the reported AppJankStat matches what was
passed in when JankTracking was first enabled. This makes sure that all
metrics and frames are associated to the app correctly.

Bug: 383586335
Test: atest CoreAppJankTestCases
Flag: EXEMPT bugfix
Change-Id: I12c827bcfc06e4e4d3de08fce0b65e5c93574c47