[Notif redesign] Center text in title-only notifs
I previously removed this centering when I was trying to fix something
for actions, but it causes a regression where the title isn't centered
vertically when a notification doesn't have a text field.

This actually only needs to NOT be centered for conversations, calls and
messages, which have an asymmetric vertical margin.

Bug: 378660052
Test: manual + screenshot tests
Flag: android.app.notifications_redesign_templates
Change-Id: Id9499f120e536ec71951deb1e3f053c83f1f7cf5