Add a function to build a header with the specified clock visibility
This CL creates a function to build an overlay shade header with the
specified clock visibility.

Ideally, we should have a stateless function for overlay shade header,
which facilitates testing. However, it is cumbersome to implement
such a stateless function, especially when some of the overlay shade
header's children accept a view model as the param. Therefore, this
function only break up the clock visibility. It is where
"PartialStateless" in the function name comes from.

BUG: 395622173
Change-Id: I0840f885fe1b332a134645f3c2ec0676017e5db8
Flag: EXEMPT refactor