Catch exceptions when comparing bundles.
We tried to catch ClassCastException in the past and log the key in
question so we could determine the root cause.  It looks like in the end
this is likely a concurrent access issue where the InCallService is
trying to modify the bundle while Telecom is updating it.  This is not
something that happens often.  Rather than taking the performance hit of
copying the Bundle in the get, or sychronizing, we'll just catch and
assume the Bundles are not equal.  Worst case we end up triggering an
extra Call.Callback#onDetailsChanged signal to the ICS.

Test: Run unit tests.
Flag: NONE Crash fix.
Fixes: 404168135
Change-Id: I5bc7a2170701fda364f6c4ca19d82a4c46745792