Add TelephonyConfigManager flag for controlling whether satellite can be disabled while enabling in progress
This CL defines a new config_support_disable_satellite_while_enable_in_progress flag which can be used to control whether the modem allows disabling satellite mode while it is enabling.

If the flag is disabled, the modem will ignore any request to disable satellite mode when it is enabling. If the flag is enabled, the modem will allow satellite mode to be disabled as long as the enable request has not been sent to the HAL.

The default value for the flag is false.

Flag: EXEMPT bugfix
Bug: 396157049
Test: SatelliteManagerTestOnMockService SatelliteControllerTest
Manual system tests

Change-Id: I80b06378b1eefbefb377f72abab39446a48bd2d6