Use ActivityClientController#enterPictureInPictureMode when restoring to WINDOWING_MODE_PINNED
When a task is restoring to pip using Activity#requestFullscreenMode, a transition is started to restore the windowing mode to pinned. Shell is currently not handling this correctly. While PipTransition is animating the task to pip, PipController#onActivityPinned is not being called and, among other things, the input consumer is not being registered.

This change uses the existing framework to request pip when attempting to restore to WINDOWING_MODE_PINNED so that the transition is correctly handled.

Flag: com.android.window.flags.enable_request_fullscreen_bugfix
Bug: 404901863
Test: Test entering fullscreen/restoring to pip using test app
Change-Id: I5a3af5cecc7623d23d13e2e39307e3787ee47b68