Fail request when fullscreen requested from split
The request fullscreen API is meant to request for the activity to take
up the full bounds of its parent. Previously, when a split task
requested fullscreen, it was reparented out of split and set to
fullscreen. This extra step does not make sense for this API and is
causing various bugs since it is not the purpose of the API.

Instead, we now check if the task is in multi_window, and if so, the app
is notified that the task is already fully expanded.

Bug: 402973271
Test: request fullscreen from split using test app
Flag: com.android.window.flags.enable_request_fullscreen_bugfix
Change-Id: I61eea857019fb5c85058bdd1b8f106eeec3f8a77