Add functionality move tasks on IME signals
With this CL we calculate the new bounds that the tasks should be moved
to when the IME starts poisitioning in desktop mode. Then we move the
tasks using a transition. The previous bounds of the tasks are stored
and when the IME stops showing the task moves back to its original
place.

Test: atest DesktopImeHandlerTest
Flag: com.android.window.flags.enable_desktop_ime_bugfix
Bug: 388570293

Change-Id: I5b8f4950dfc2a829095bc91d586d11c503746b02